requirements/requirements.txt
# Utilities
yacs==0.1.8
omegaconf==2.3.0
PyYAML==6.0.2
pandas==2.2.3
tqdm==4.67.1
tabulate==0.9.0
scikit-learn==1.6.1

# CV + data
opencv-python==4.10.0.84
scikit-image==0.24.0
albumentations==1.4.24
Pillow==11.1.0
h5py==3.11.0

# PRTReid deps
Cython==3.0.12
deepdiff==8.6.1
gdown==5.2.0
wandb==0.19.8
monai==1.4.0
torchmetrics==0.10.3
